stream_base,tls_wrap: notify on destruct
Classes like TLSWrap hold a pointer to another StreamBase instance. The
lifetime of these two are independent and allowed one to be retained
while the other is cleaned up. In the case of TLSWrap, the private
member TLSWrap::stream_ would be deleted, while the TLSWrap could still
attempt to access it to see if it was alive.

Give StreamBase a destruct callback to notify when the class is being
cleaned up. Use this in TLSWrap to clear out the field.